# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "doc-processing",
    "description": "Document processing, conversion to markdown, chunking for RAG, and LLM completions.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/documents/convert-ms-docs": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Convert Ms Docs",
        "description": "Convert a Microsoft Office document (Word/Excel/PowerPoint) to markdown from file upload or URL.",
        "operationId": "convert_ms_docs_documents_convert_ms_docs_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_ms_docs_documents_convert_ms_docs_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/process-using-unstructured": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Process Using Unstructured",
        "description": "Process a PDF or image using Unstructured: extract tables, images, and metadata.\n\nAccepts either an uploaded file or a URL.",
        "operationId": "process_using_unstructured_documents_process_using_unstructured_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          },
          {
            "name": "file_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pdf",
                    "image",
                    "markdown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Type"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_using_unstructured_documents_process_using_unstructured_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessFromUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/pdf-to-markdown": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Pdf To Markdown Endpoint",
        "description": "Convert a PDF to markdown using Docling (easyocr or vlm backend) from file upload or URL.",
        "operationId": "pdf_to_markdown_endpoint_documents_pdf_to_markdown_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          },
          {
            "name": "backend",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "easyocr",
                "vlm"
              ],
              "type": "string",
              "default": "easyocr",
              "title": "Backend"
            }
          },
          {
            "name": "merge_tables",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Merge Tables"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pdf_to_markdown_endpoint_documents_pdf_to_markdown_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfToMarkdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/xbrl-to-markdown": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Xbrl To Markdown Endpoint",
        "description": "Convert an XBRL instance document (.xml/.xbrl) to markdown using Docling's XBRL backend from file or URL.",
        "operationId": "xbrl_to_markdown_endpoint_documents_xbrl_to_markdown_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_xbrl_to_markdown_endpoint_documents_xbrl_to_markdown_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XbrlToMarkdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/process-using-deepseek": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Process Using Deepseek",
        "description": "Convert a PDF or image to markdown using DeepSeek-OCR via Ollama.\n\nAccepts either an uploaded file or a URL. Only PDF and common image types are supported.",
        "operationId": "process_using_deepseek_documents_process_using_deepseek_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_using_deepseek_documents_process_using_deepseek_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfToMarkdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/process-using-glm": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Process Using Glm",
        "description": "Convert a PDF or image to markdown using GLM-OCR via Ollama.\n\nAccepts either an uploaded file or a URL. Only PDF and common image types are supported.",
        "operationId": "process_using_glm_documents_process_using_glm_post",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "minLength": 1,
                  "maxLength": 2083
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_using_glm_documents_process_using_glm_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfToMarkdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/chunk": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Chunk Document",
        "description": "Split text into overlapping chunks for RAG.",
        "operationId": "chunk_document_documents_chunk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_convert_ms_docs_documents_convert_ms_docs_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_convert_ms_docs_documents_convert_ms_docs_post"
      },
      "Body_pdf_to_markdown_endpoint_documents_pdf_to_markdown_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_pdf_to_markdown_endpoint_documents_pdf_to_markdown_post"
      },
...
      "Body_process_using_unstructured_documents_process_using_unstructured_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_process_using_unstructured_documents_process_using_unstructured_post"
      },
...
      "ConvertResponse": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown",
            "description": "Converted markdown content"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "ConvertResponse",
        "description": "Result of document-to-markdown conversion."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PdfToMarkdownResponse": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown",
            "description": "Converted markdown content"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "PdfToMarkdownResponse",
        "description": "Result of PDF-to-markdown conversion."
      },
      "ProcessFromUrlResponse": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableItem"
            },
            "type": "array",
            "title": "Tables",
            "description": "Extracted tables"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageItem"
            },
            "type": "array",
            "title": "Images",
            "description": "Extracted images"
          },
          "extract_output_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extract Output Dir",
            "description": "Dir path if images saved to disk"
          }
        },
        "type": "object",
        "required": [
          "tables",
          "images"
        ],
        "title": "ProcessFromUrlResponse",
        "description": "Result of process-from-url: tables and images."
      },
...
      "XbrlToMarkdownResponse": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown",
            "description": "Converted markdown content"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "XbrlToMarkdownResponse",
        "description": "Result of XBRL-to-markdown conversion."
      }
    }
  }
}